fix(connect): cast bookings.price::numeric ใน commission webhook 1.8.3+114
bookings.price เป็น double precision แต่ connect_merchant_gp_amount รับ numeric
Postgres ไม่ cast double precision -> numeric แบบ implicit ตอน resolve function
ทำให้ trigger notify_storeos_order / notify_storeos_order_created error
'function does not exist' ตอนรันจริง (มี active StoreOS connection) -> การ
เปลี่ยนสถานะ/สร้างออเดอร์อาหารพัง. แก้โดย cast NEW.price::numeric /
v_booking.price::numeric ที่จุดเรียก + smoke test ล็อก cast.

Verify (prod, real data): 55 -> commission 14 -> net 41 (ตรงเคสในตั๋ว),
65->11, 120->13, 60->10; node --test -> 21 pass. Migration re-applied to prod,
edge function connect_update_order_status deployed.

+++ supabase/migrations/20260704090000_connect_food_ready_origin_and_commission.sql
@@ -238,7 +238,8 @@ BEGIN
       -- price = ยอดอาหาร (ไม่รวมค่าส่ง), commission = GP ที่หักจากยอดอาหาร
       -- StoreOS แสดงยอดจ่ายร้าน = merchant_total - commission
       'merchant_total', v_booking.price,
-      'commission', public.connect_merchant_gp_amount(v_booking.merchant_id, v_booking.price),
+      -- price เป็น double precision — ต้อง cast เป็น numeric ให้ตรง signature ของ fn
+      'commission', public.connect_merchant_gp_amount(v_booking.merchant_id, v_booking.price::numeric),
       'total', coalesce(v_booking.price, 0) + coalesce(v_booking.delivery_fee, 0),
       'dropoff', jsonb_build_object('address', v_booking.destination_address),
       'paid', true,
@@ -313,7 +314,8 @@ BEGIN
     'status', NEW.status,
     'total', coalesce(NEW.price, 0) + coalesce(NEW.delivery_fee, 0),
     'merchant_total', NEW.price,
-    'commission', public.connect_merchant_gp_amount(NEW.merchant_id, NEW.price),
+    -- price เป็น double precision — ต้อง cast เป็น numeric ให้ตรง signature ของ fn
+    'commission', public.connect_merchant_gp_amount(NEW.merchant_id, NEW.price::numeric),
     'paid', true,
     'ts', extract(epoch FROM now())::bigint
   );
